requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
django-tables2==2.7.0
sqlparse==0.5.1
typing_extensions==4.12.2
